Rest/USAS/requirements.txt
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
colorama==0.3.9
decorator==4.0.11
Django==1.11.1
django-crispy-forms==1.6.1
django-registration-redux==1.6
djangorestframework==3.6.3
enum34==1.1.6
ipdb==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
MySQL-python==1.2.5
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
Pygments==2.2.0
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
svgwrite==1.1.10
traitlets==4.3.2
wcwidth==0.1.7
win-unicode-console==0.5
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
colorama==0.3.9
decorator==4.0.11
defusedxml==0.5.0
Django==1.11.1
django-crispy-forms==1.6.1
django-registration-redux==1.6
djangorestframework==3.6.3
enum34==1.1.6
ipdb==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
jira==1.0.10
MySQL-python==1.2.5
oauthlib==2.0.2
olefile==0.44
ordereddict==1.1
packaging==16.8
pathlib2==2.2.1
pbr==3.0.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
Pygments==2.2.0
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
svgwrite==1.1.10
traitlets==4.3.2
wcwidth==0.1.7
win-unicode-console==0.5
robotframework==3.0.2

